+++ src/Command/ClearAllCommand.php
@@ -68,6 +68,6 @@ class ClearAllCommand extends Command
             $this->abort();
         }
 
-        $io->verbose(__d('thumber', 'Thumbnails deleted: {0}', $count));
+        $io->verbose(__d('thumber', 'Thumbnails deleted: {0}', $count ?? 0));
     }
 }