+++ tests/TestCase/View/Helper/ThumbHelperTest.php
@@ -16,7 +16,6 @@ namespace Thumber\Cake\Test\TestCase\View\Helper;
 
 use BadMethodCallException;
 use Cake\View\View;
-use Intervention\Image\Exception\InvalidArgumentException as InterventionInvalidArgumentException;
 use InvalidArgumentException;
 use Thumber\Cake\TestSuite\TestCase;
 use Thumber\Cake\View\Helper\ThumbHelper;
@@ -27,6 +26,11 @@ use Thumber\Cake\View\Helper\ThumbHelper;
 class ThumbHelperTest extends TestCase
 {
     /**
+     * @var \Thumber\Cake\View\Helper\ThumbHelper
+     */
+    protected $Thumb;
+
+    /**
      * Called before every test method
      * @return void
      */
@@ -34,14 +38,14 @@ class ThumbHelperTest extends TestCase
     {
         parent::setUp();
 
-        $this->Thumb = new ThumbHelper(new View());
+        $this->Thumb = $this->Thumb ?: new ThumbHelper(new View());
     }
 
     /**
      * Test for magic `__call()` method
      * @test
      */
-    public function testMagicCall()
+    public function testMagicCall(): void
     {
         $path = '400x400.png';
         $params = ['width' => 200];
@@ -75,6 +79,7 @@ class ThumbHelperTest extends TestCase
         //Calling a no existing method
         $this->expectException(BadMethodCallException::class);
         $this->expectExceptionMessage('Method `Thumber\Cake\Utility\ThumbCreator::noExisting()` does not exist');
+        /** @phpstan-ignore-next-line */
         $this->Thumb->noExisting('400x400.png');
     }
 
@@ -82,9 +87,9 @@ class ThumbHelperTest extends TestCase
      * Test for magic `_call()` method, called without parameters
      * @test
      */
-    public function testMagicCallWithoutParameters()
+    public function testMagicCallWithoutParameters(): void
     {
-        $this->expectException(InterventionInvalidArgumentException::class);
+        $this->expectException(InvalidArgumentException::class);
         $this->Thumb->crop('400x400.png');
     }
 
@@ -92,10 +97,11 @@ class ThumbHelperTest extends TestCase
      * Test for magic `_call()` method, called without path
      * @test
      */
-    public function testMagicCallWithoutPath()
+    public function testMagicCallWithoutPath(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('Thumbnail path is missing');
+        /** @phpstan-ignore-next-line */
         $this->Thumb->crop();
     }
 
@@ -103,7 +109,7 @@ class ThumbHelperTest extends TestCase
      * Test for magic `isUrlMethod()` method
      * @test
      */
-    public function testIsUrlMethod()
+    public function testIsUrlMethod(): void
     {
         $isUrlMethod = function (string $methodName) {
             return $this->invokeMethod($this->Thumb, 'isUrlMethod', [$methodName]);